feat: ios features extraction
