+++ ios/AudioProcessingHelpers.swift
@@ -1,0 +1,102 @@
+// AudioProcessingHelpers.swift
+
+import Accelerate
+
+func extractMFCC(from segment: [Float], sampleRate: Float) -> [Float] {
+    // Placeholder for MFCC extraction logic
+    return []
+}
+
+func extractSpectralCentroid(from segment: [Float], sampleRate: Float) -> Float {
+    Logger.debug("Extracting Spectral Centroid from segment of length \(segment.count)")
+    
+    let length = segment.count
+    let log2n = UInt(round(log2(Double(length))))
+    
+    guard let fftSetup = vDSP_create_fftsetup(log2n, Int32(kFFTRadix2)) else {
+        Logger.debug("Failed to create FFT setup")
+        return 0.0
+    }
+    
+    var realp = [Float](repeating: 0, count: length / 2)
+    var imagp = [Float](repeating: 0, count: length / 2)
+    var magnitudes = [Float](repeating: 0.0, count: length / 2)
+    var spectralCentroid: Float = 0.0
+    
+    segment.withUnsafeBufferPointer { bufferPointer in
+        realp.withUnsafeMutableBufferPointer { realpPtr in
+            imagp.withUnsafeMutableBufferPointer { imagpPtr in
+                var splitComplex = DSPSplitComplex(realp: realpPtr.baseAddress!, imagp: imagpPtr.baseAddress!)
+                
+                bufferPointer.baseAddress!.withMemoryRebound(to: DSPComplex.self, capacity: length / 2) { typeConvertedTransferBuffer in
+                    vDSP_ctoz(typeConvertedTransferBuffer, 2, &splitComplex, 1, vDSP_Length(length / 2))
+                }
+                
+                vDSP_fft_zrip(fftSetup, &splitComplex, 1, log2n, Int32(FFT_FORWARD))
+                
+                vDSP_zvmags(&splitComplex, 1, &magnitudes, 1, vDSP_Length(length / 2))
+                
+                var totalMagnitude: Float = 0.0
+                var weightedSum: Float = 0.0
+                for i in 0..<magnitudes.count {
+                    let magnitude = magnitudes[i]
+                    totalMagnitude += magnitude
+                    weightedSum += Float(i) * magnitude
+                }
+                
+                spectralCentroid = totalMagnitude > 0 ? weightedSum / totalMagnitude : 0.0
+                Logger.debug("Spectral Centroid: \(spectralCentroid)")
+            }
+        }
+    }
+    
+    vDSP_destroy_fftsetup(fftSetup)
+    
+    return spectralCentroid
+}
+
+func extractSpectralFlatness(from segment: [Float]) -> Float {
+    Logger.debug("Extracting Spectral Flatness from segment of length \(segment.count)")
+    
+    var mean: Float = 0.0
+    var geometricMean: Float = 1.0
+    let count = vDSP_Length(segment.count)
+    
+    vDSP_meamgv(segment, 1, &mean, count)
+    
+    var sumLogValues: Float = 0.0
+    for value in segment {
+        let adjustedValue = max(value, 1e-10)
+        sumLogValues += log(adjustedValue)
+    }
+    geometricMean = exp(sumLogValues / Float(count))
+    
+    let spectralFlatness = mean > 0 ? geometricMean / mean : 0.0
+    Logger.debug("Spectral Flatness: \(spectralFlatness)")
+    return spectralFlatness
+}
+
+func extractSpectralRollOff(from segment: [Float], sampleRate: Float) -> Float {
+    // Implement spectral roll-off extraction logic
+    return 0.0
+}
+
+func extractSpectralBandwidth(from segment: [Float], sampleRate: Float) -> Float {
+    // Implement spectral bandwidth extraction logic
+    return 0.0
+}
+
+func extractChromagram(from segment: [Float], sampleRate: Float) -> [Float] {
+    // Implement chromagram extraction logic
+    return []
+}
+
+func extractTempo(from segment: [Float], sampleRate: Float) -> Float {
+    // Implement tempo extraction logic
+    return 0.0
+}
+
+func extractHNR(from segment: [Float]) -> Float {
+    // Implement harmonic-to-noise ratio extraction logic
+    return 0.0
+}
