+++ ios/ExpoAudioStreamModule.swift
@@ -17,6 +17,15 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
             streamManager.delegate = self
         }
         
+        /// Asynchronously extracts audio analysis data from an audio file.
+        ///
+        /// - Parameters:
+        ///   - options: A dictionary containing:
+        ///     - `fileUri`: The URI of the audio file.
+        ///     - `pointsPerSecond`: The number of data points to extract per second of audio.
+        ///     - `algorithm`: The algorithm to use for extraction.
+        ///     - Additional optional parameters to specify which features to extract (e.g., `energy`, `mfcc`, `rms`, etc.).
+        ///   - promise: A promise to resolve with the extracted audio analysis data or reject with an error.
         AsyncFunction("extractAudioAnalysis") { (options: [String: Any], promise: Promise) in
             guard let fileUri = options["fileUri"] as? String,
                   let url = URL(string: fileUri),
@@ -26,10 +35,10 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
                 return
             }
             
-            // Perform audio analysis on a background thread
+            let featureOptions = self.extractFeatureOptions(from: options)
+            
             DispatchQueue.global().async {
-                if let result = extractAudioAnalysis(fileUrl: url, pointsPerSecond: pointsPerSecond, algorithm: algorithm) {
-                    
+                if let result = extractAudioAnalysis(fileUrl: url, pointsPerSecond: pointsPerSecond, algorithm: algorithm, featureOptions: featureOptions) {
                     Logger.debug("Extraction result: \(result)")
                     let resultDict = result.toDictionary()
                     promise.resolve(resultDict)
@@ -39,6 +48,54 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
             }
         }
         
+        /// Asynchronously extracts waveform data from an audio file.
+        ///
+        /// - Parameters:
+        ///   - options: A dictionary containing:
+        ///     - `fileUri`: The URI of the audio file.
+        ///     - `numberOfSamples`: The number of waveform samples to extract.
+        ///     - `offset`: The offset (in samples) to start reading from (optional, default is 0).
+        ///     - `length`: The length (in samples) of the audio to read (optional).
+        ///   - promise: A promise to resolve with the extracted waveform data or reject with an error.
+        AsyncFunction("extractWaveform") { (options: [String: Any], promise: Promise) in
+            guard let fileUri = options["fileUri"] as? String,
+                  let url = URL(string: fileUri),
+                  let numberOfSamples = options["numberOfSamples"] as? Int else {
+                promise.reject("INVALID_ARGUMENTS", "Invalid arguments provided")
+                return
+            }
+            
+            let offset = options["offset"] as? Int ?? 0
+            let length = options["length"] as? UInt
+            
+            DispatchQueue.global().async {
+                do {
+                    let extractor = try WaveformExtractor(url: url, resolve: { result in
+                        promise.resolve(result)
+                    }, reject: { code, message in
+                        promise.reject(code, message)
+                    })
+                    if let waveform = extractor.extractWaveform(numberOfSamples: numberOfSamples, offset: offset, length: length) {
+                        promise.resolve(waveform)
+                    } else {
+                        promise.reject("EXTRACTION_ERROR", "Failed to extract waveform")
+                    }
+                } catch {
+                    promise.reject("EXTRACTION_ERROR", "Failed to initialize waveform extractor")
+                }
+            }
+        }
+        
+        
+        /// Asynchronously starts audio recording with the given settings.
+        ///
+        /// - Parameters:
+        ///   - options: A dictionary containing:
+        ///     - `sampleRate`: The sample rate for recording (default is 16000.0).
+        ///     - `channelConfig`: The number of channels (default is 1 for mono).
+        ///     - `audioFormat`: The bit depth for recording (default is 16 bits).
+        ///     - `interval`: The interval in milliseconds at which to emit recording data (default is 1000 ms).
+        ///   - promise: A promise to resolve with the recording settings or reject with an error.
         AsyncFunction("startRecording") { (options: [String: Any], promise: Promise) in
             self.checkMicrophonePermission { granted in
                 guard granted else {
@@ -68,10 +125,17 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
             }
         }
         
+        /// Retrieves the current status of the audio stream.
+        ///
+        /// - Returns: The current status of the audio stream.Ï
         Function("status") {
             return self.streamManager.getStatus()
         }
         
+        /// Asynchronously stops audio recording and retrieves the recording result.
+        ///
+        /// - Parameters:
+        ///   - promise: A promise to resolve with the recording result or reject with an error.
         AsyncFunction("stopRecording") { (promise: Promise) in
             if let recordingResult = self.streamManager.stopRecording() {
                 // Convert RecordingResult to a dictionary
@@ -90,16 +154,29 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
             }
         }
         
+        /// Asynchronously lists all audio files stored in the document directory.
+        ///
+        /// - Parameters:
+        ///   - promise: A promise to resolve with the list of audio file URIs or reject with an error.
+        /// - Returns: A promise that resolves with the list of audio file URIs or rejects with an error.
         AsyncFunction("listAudioFiles") { (promise: Promise) in
             let files = listAudioFiles()
             promise.resolve(files)
         }
         
+        /// Clears all audio files stored in the document directory.
         Function("clearAudioFiles") {
             clearAudioFiles()
         }
     }
     
+    /// Handles the reception of audio data from the AudioStreamManager.
+    ///
+    /// - Parameters:
+    ///   - manager: The AudioStreamManager instance.
+    ///   - data: The received audio data.
+    ///   - recordingTime: The current recording time.
+    ///   - totalDataSize: The total size of the received audio data.
     func audioStreamManager(_ manager: AudioStreamManager, didReceiveAudioData data: Data, recordingTime: TimeInterval, totalDataSize: Int64) {
         guard let fileURL = manager.recordingFileURL,
               let settings = manager.recordingSettings else { return }
@@ -132,6 +209,10 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
         sendEvent(audioDataEvent, eventBody)
     }
     
+    /// Checks microphone permission and calls the completion handler with the result.
+    ///
+    /// - Parameters:
+    ///   - completion: A completion handler that receives a boolean indicating whether the microphone permission was granted.
     private func checkMicrophonePermission(completion: @escaping (Bool) -> Void) {
         switch AVAudioSession.sharedInstance().recordPermission {
         case .granted:
@@ -149,6 +230,7 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
         }
     }
     
+    /// Clears all audio files stored in the document directory.
     private func clearAudioFiles() {
         let fileURLs = listAudioFiles()  // This now returns full URLs as strings
         fileURLs.forEach { fileURLString in
@@ -163,9 +245,33 @@ public class ExpoAudioStreamModule: Module, AudioStreamManagerDelegate {
                 print("Invalid URL string: \(fileURLString)")
             }
         }
-        
     }
     
+    /// Extracts feature options from the provided options dictionary.
+    ///
+    /// - Parameters:
+    ///   - options: The options dictionary containing feature flags.
+    /// - Returns: A dictionary with feature flags and their boolean values.
+    private func extractFeatureOptions(from options: [String: Any]) -> [String: Bool] {
+        return [
+            "energy": options["energy"] as? Bool ?? false,
+            "mfcc": options["mfcc"] as? Bool ?? false,
+            "rms": options["rms"] as? Bool ?? false,
+            "zcr": options["zcr"] as? Bool ?? false,
+            "dB": options["dB"] as? Bool ?? false,
+            "spectralCentroid": options["spectralCentroid"] as? Bool ?? false,
+            "spectralFlatness": options["spectralFlatness"] as? Bool ?? false,
+            "spectralRollOff": options["spectralRollOff"] as? Bool ?? false,
+            "spectralBandwidth": options["spectralBandwidth"] as? Bool ?? false,
+            "chromagram": options["chromagram"] as? Bool ?? false,
+            "tempo": options["tempo"] as? Bool ?? false,
+            "hnr": options["hnr"] as? Bool ?? false
+        ]
+    }
+    
+    /// Lists all audio files stored in the document directory.
+    ///
+    /// - Returns: An array of file URIs as strings.
     func listAudioFiles() -> [String] {
         guard let documentDirectory = try? FileManager.default.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: false) else {
             print("Failed to access document directory.")
